implement user location updates and continual plotting
At each time step in run(), loop over users and increment their
locations. Do this by finding which path they appear to be on, getting
the next waypoint on that path, and interpolating along the line
connecting their current true position to the waypoint. (Noisy. So
there's no noise in how the true positions are updated (yet), but the
observed locations have noise added.